refactor: gather all distance construction into one build module
One symmetric 2x2 of fills — {condensed, matrix} x {sequential, parallel} — behind two entry points that own validation, cosine normalization, and the parallel switch. The five per-metric sequential condensed fills collapse into the same metric-branching shape the other fills use; _compute keeps pair arithmetic, _store keeps storage and reads. Bit-identity of the reshape is pinned by the cross-backend canary against the untouched on-demand reads.
